backend/requirements.txt
Flask==2.2.2
Flask_Cors==3.0.10
Flask_Mail==0.9.1
imbalanced_learn==0.9.1
imblearn==0.0
matplotlib==3.6.2
nltk==3.7
numpy==1.23.4
pandas==1.5.2
PyJWT==2.6.0
scikit_learn==1.1.3
seaborn==0.12.1
tweepy==4.12.1
wordcloud==1.8.2.2
